Assets, Movement update, "mob"
I updated the movement to make it more planar instead of platformer. I also added some placeholder assets for when I flesh out the code and can afford to make better looking assets. Finally I also made a "mob" that follows the character as any given position on the board.